♻️ `assistantMessageHandler`をファイル分割した
